-enh restructured sensor_base and sensor_helper
